requirements.txt
# === Core (MVP) ===
pymupdf==1.24.10 
rapidfuzz==3.9.6   
pydantic==2.9.2      

python-dotenv
openai>=1.0.0

fastapi
uvicorn[standard]

python-multipart==0.0.9
